Garbage in, rainbows out
This commit includes a lot of changes to how problematic cases are handled.
The basic idea is that first the code snippet is tested on good and bad
revisions (git is not doing that by default), and further behavior is adapted
according to that.

Also, 2015.12 is now used instead of v6.c.