Disable colors in error messages
We can replace terminal colors with something sensible, but the problem is that
it is not going to work so well for gists. At the time we execute something,
there is no way to know if we are going to gist it or not. It seems like
there's no easy way to have a cake and eat it.