Bye-bye Perl 5 bots
We strive to increase the amount of dog food in our diet, so Perl 5 variants of
the bots have to go.

Perl 6 bots are mostly identical. They have some new features that previous
bots did not have, and do not have some bugs that were fixed during the rewrite.
They are also more async and the code is more beautiful.

That being said, they are also more unstable… we will have to deal with it.