Compress builds to save space
Also, bash scripts were replaced with perl 6 code.

zstd was chosen for its fast decompression speed and good compression ratio. See
issue #23 for more info.

Rakudo is now being installed into “/tmp”. This way it will be easier for
others to extract archives on their system if we ever publish them (because
/tmp is writeable on almost any system). For example, we can let people run
git bisect locally by downloading prepared builds.

This means that all previous builds have to be rebuilt. That's not a big deal,
but it will take up to three days to process.

Another side effect of these changes is that bots will not see new commits
until they are actually built. Now when running something on HEAD you can be
sure that the actual commit that is pointed on by HEAD is actually ready to
roll. No more “whoops, try using some older commit” error messages.

Currently it is very unstable and more testing is required.