Decompress builds when required
This works great with Committable, but does not work with Bisectable.
Before 540aba62 bisection process was done by using a separate bash script, so
it was not possible to use run-snippet helper function. Now that we have full
control over the process, we *should* use run-snippet (which will automatically
decompresses the build as needed). Consistent use of run-snippet will unbreak
Bisectable (which was broken after huge changes in f7bd3c3b).