Add a slight delay before returning a value
Otherwise messages may be out of order. This workaround is a bit ugly, but it
will do the job for now.

Out-of-order behavior is consistent on a local server when running tests, but
on a real server it is less common (but still observable).

It is likely that IRC::Client should provide something by default to prevent
this from happening.