Fix occasional crashes associtated with timeouts
Just sending a SIGHUP does not mean that we can close the channel, the
process is probably still alive and it will attempt to send stuff into
a closed channel.

This fixes #60. Previously this was the cause of MoarVM panics (until
it was fixed in Rakudo), but generally this is our fault.

However, there is still at least one bug in rakudo, which is why we
did not notice this problem right away. When the command prints a lot
of stuff into stdin, it will never time out. The whole thing is stuck
in a Proc::Async tap, without giving any chance for Promice.in(…) to
fire up.