IRC colors and factor out of run-smth
Now all bots should have colorful output in IRC while still having
normal gists (everything is going to have eject symbol now).

Tests with control characters are disabled for now.

Refactoring of run-snippet is needed because we might want to do other
stuff with the build (this will be used in satisfiable).