Cover the whole unicode range
While 0..0x1FFFF was a nice approximation, nowadays unicodable is fast
enough to go through the whole range (0..0x10FFFF) in a reasonable
amount of time.

Code blocks are not guaranteed to run as fast (the user might include
whatever he wants in them), therefore the timeout had to be bumped.

There are ways we can optimize it, but it is good enough for now.