Make Committable friendlier
Sometimes people forget to specify a commit. Usually this would mean
that the commit will be parsed as “say”, so let's use “6c” instead and
put “say” back into the code. If the snippet does not start with
“say”, it will simply complain that it cannot find this commit.

This commit also adds a “Did you mean “…”?” suggestion. This
suggestion includes some custom options, all tags for which a build
exists, and all commit hashes since 2014-01-01. The latter introduces
a 1-3 second delay (depending on the input), but that's OK. This delay
is only there if you did not provide a right revision name, and it is
is not noticeable in practice. Still, I am hoping that rakudo will
become faster over time and this delay will diminish completely.

Addresses issues #73 #85