Test the fallback response
Adds a test for issue #67.

Generally, this fallback was added to handle buggy behavior. Any bot
should figure out what to respond for *any* given command. Committable
is the only bot that does not do so, which is why we are able to test
the fallback response. Arguably, this should be changed, but the
default response is actually spot on in this particular case.