Commit build.p6 as is
For a long time this script was suffering due to various instabilities
in rakudo (see issue #24). As of today, some of the bugs were fixed,
but at the same time this script itself was wiggled into some stable
state. I am afraid to touch it.

Therefore, committing what we have without any clean up.

In fact, there are some issues I can see right now, but the script is
proven to work in practice… 🙈