Detect messages starting with “say” and answer sometimes
This change has been there for a while, I simply forgot to commit
it. It seems that there was at least one time it answered
unnecessarily, so perhaps the regex can be improved.

But generally, the idea is really nice. If it becomes annoyng for some
reason, I can always make the regex more strict.