Don't build everything in one go
In some cases the build script may focus on something
irrelevant (e.g. old commits or commits in a different branch), which
means that whatever is committed during this process will not be built
until everything is done. By adding a limit we ensure that the process
ends quicker, and when it is restarted it will first work on the most
recent commits.