Yet another commit range
You'd say “Alex, but why can't we just extend COMMIT-RANGE so that it
starts from 2014.01?”. I'll tell you.

The server is not all-powerful and all-knowing. Therefore, we must pay
attention to the order of commits to get best results.

This way, for the normal range, we want to start building from the most
recent commit (so that the latest stuff becomes available sooner), and
the rest can catch up later. So if somebody pushed 10 commits to the
repo, the last one will become available sooner. Note, however, that
most bots will not know about it because the repo used for getting the
information is only updated when the build script exits, but you can
force bots to use the required build by telling them the sha directly.

However, if you want to build ancient pyramids, you cannot use the same
approach. In a long term you will run into space issues, so there are
ways to optimize disk usage by putting more than one build into an
archive. It is logical that for best results you'd want adjacent commits
to be in the same archive, but some things get more complicated if you
want to do the build process backwards. Therefore, for old stuff, we
start from the oldest commit (this is not strictly obligatory, but it
just makes things easier).