Long-term storage using lrzip (issue #117)
After all of the experiments discussed in #117, it was decided to use 20
builds per archive with --level=9, all other settings being
default. This has freed a lot of space on the server, and the introduced
delay for using older builds does not seem to be high at all (≈2
seconds). Of course, tagged commits are not stored this way, so using
2014.01 or other builds like that is still as fast, but 2014.02^ is a
bit slower. I wonder if anybody will ever notice this.