The absence of -- in a command is usually a bug
In this case, a query like “HEAD grep=-V say 42” was failing. Not that
it is a sensible query, but it shouldn't fail like that.

Another possible solution is to use ‘-e’ (which expects a pattern in the
next argument), but it is slightly less explicit as I'm used to seeing
-- everywhere (and also assuming that there's a bug if I don't see it).