Gist exceptions (resolves issue #106)
Whenever some exception was thrown, bots used to simply ignore what
happened and not report back with anything. With this commit not only
we catch exceptions, but also:
* Gist them and provide a link.
* Notify maintainers on #whateverable channel.
* Files in the exception are linked to actual files in the repo
  (directly to the offending line).
* If the file is not committed for some reason, we gist the file as
  well. Links in the exception will point to local files of the gist.
* The output of ｢git diff HEAD｣ is provided so that it is clear what
  changes are present on the server. Normally users will not need
  this, but this is useful for development.

Example:

  <AlexDaniel> grep: kill[^-']
  <greppable6test> AlexDaniel, No! It wasn't me! It was the one-armed man! Backtrace: https://gist.github.com/09f1f2d9b04ce5a84df66eee787a16d4

And this appears on #whateverable channel:

  <greppable6test> AlexDaniel, MasterDuke: I'm acting stupid on #whateverable. Help me.

This commit also resolves issue #162, but in a slightly different
way. As we are using Proc::Async, there is no :merge, you just do
｢whenever $proc｣. However, it does not seem to work reliably (yet), so
we are not doing it. In any case, the code is much simpler now.

Issue #153 was partially addressed. The exception was happening on a
different thread, and therefore was not caught by any CATCH
block. Because of that, the bot used to crash without a notice. Now
that the get-output method is rewritten with react/whenever, the
exception does not really happen on a different thread, and therefore
the situation is handled much better.

Also, “beg-for-help” method was removed, as I believe it is not needed
anymore. Current exception handling is superior to what we had before.

What happens if an exception is thrown while we are handling some
other exception? Well, we print a different message for it.

The situation is much more complicated in filters, because not only
the exception handling can fail, but *the result* of it may explode
later in the filter. Therefore, there's a third message for this
situation.

And that's pretty much it. This may seem as little work, but it brings
the bots onto a whole different level.