Show all characters if possible
Unicodable will now print all characters in its last message (only if
it makes sense to do so).

Resolves issue #104.

Note that with these changes it may call get-description method more
than once for some characters (up to two extra calls in total!), but
this really does not matter. What matters, however, is that now it
will respond faster if it goes beyond $LIMIT (because it does not have
to process characters early).

It turns out that checking if it is safe to print all characters is
difficult, but not too much. If it round trips nicely (e.g. no
normalization taking effect), and if GCB property tells us there are
no weirdos, then it is probably safe to concat the whole thing and
print it. Hopefully parens around it won't affect anything.