Introduce sub grumble
This will replace almost all instances of “die” in the code. I am not
entirely convinced that this is the best approach, but the idea is to
have some sort of soft exceptions that are simply replied back to the
user (e.g. provided url is invalid).

The reason I'm doing this is because the code was very dirty in many
cases due to handling of unwanted results. By having a mechanism that
bails out (no matter how deep it is) we no longer need to have a bunch
of checks everywhere.

Right now it allows to remove a lot of annoying code. If there is any
better way, I'm ready to improve it further :)