Add Reply to any reply…
Every bot had code that wrapped around the whole logic and had to mix
in Reply in order to have everything working properly. Given that now
we wrap around everything anyway, just do it there.

Note that it's OK to have Nil there, as it simply indicates that
nothing should be sent (and if so, we don't have to add Reply to this
Nil…).

Actual changes to the bots will be committed shortly in huge
refactoring commits.