Let TWEAK be a method
It does not work anyway, so let's say it's just a method.

The problem is that we want TWEAK to fire every time an object is
created, but as we have it in a role, it only works if the class does
not have its own TWEAK method. Quite possibly this is a rakudobug, but
I am not sure.