Refactoring of tests
A lot of simple tests were repeated for each bot, so now they are part
of Testable. This also allowed to easily add a couple of other small
tests.

Unfortunately, we still need to do one extra test at the end of every
file, which is something Testable should do for us by
itself. Hopefully this will be addressed later.