Generate the last test automatically
The problem that these tests are trying to solve is probably
non-obvious, so here's the explanation. The test bench works by
sending messages to the bot and waiting for responses. Let's say you
have a test that expects two messages (‘foo’ and ‘bar’), but the bot
replies with three instead (‘foo’, ‘bar’, and 3 seconds later ‘baz’)?

This behavior will only be caught in the next test (where this extra
message will be treated by the test bench as the first response). This
is LTA because the failing test will not be the one that is actually
failing…  but as long as at least something is “not ok”, we can work
with it. After all, system testing will never tell you what failed
exactly.

There is no problem as long as the test is not last in the file
(obviously there is always one test that is last). This is why I used
to put an extra test (normally a copy of the first test) to ensure
that the last test is totally OK.

This gets annoying rather quickly, so now there is an easy way to
rerun the first test.