Longer lines are easier to read
Previous limit was too short. I think the idea was to have different
behavior depending on the data: one for printing the result on the
same line (“horizontal”?), and another for doing what it does now
(vertical!). I think it's better to abandon the idea so that we have
consistent output (even though it may be LTA in some cases).