Kill everybody before upgrading rakudo
There are so many things that can be done better here (e.g. the build
script may be running and so we should terminate it gracefully), but
that's a quest for another day.

Also, this does not work anyway. Why? Rakudobrew pulls rakudo from
under the sake script, so it dies a horrible death.

💩