Everyone loves performance improvements! №2
See also f2d1d8c13e673da1080216873a70467062f8668f.

You may think that tests with unbounded queries are taking more time
because we simply have more commits now, but this does not matter
given the logarithmic nature of the process. Also, even if we have one
more commit to test now, this does not explain why things were working
properly before I upgraded rakudo on the server.

To me it seems that the bot is getting slower over time. You may see
that I touched the first test in this commit, *but this is because the
first test is implicitly repeated at the end of every test file*. So
what actually happens is that this test passes just fine at the
beginning, but times out at the end. This tells me that there is some
performance degradation of long-running processes, but I can't tell
much right now (not enough data).

I think this commit and f2d1d8c13e673da1080216873a70467062f8668f
should be reverted once the issue is gone.