New bot: Releasable (with tests)
Automatically detects release dates, commits that were logged, etc.
Also, prints warnings if something is not quite right.

Technically, “changelogable” would be more precise (given that the bot
does not make releases itself), but “releasable” is easier to remember.

The data is taken from the rakudo repo directly (docs/ChangeLog,
docs/release_guide.pod, VERSION). This means that no special actions
should be taken to keep the bot up-to-date. Anyone can edit the
changelog and the bot will pick up these changes automatically.

Commits that should not be logged must be mentioned in one of
the commits to the changelog file. This is possibly LTA, but we will
see how it goes.

Generally, this is an experiment in making the release process
more awesome. Whether it is going to work better or not, I don't
know. In any case, this is an independent bot that can be used in
addition to anything else.

There is a dependency on R6 to check for release blockers. As of
today, there are no plans to change that (mainly because the whitelist
approach of R6 is pretty much flawless).

One of the cool features of releasable is that you can give it a link
to the modified changelog file, and it will give you the information
about it as if it was in rakudo repo (e.g. commits that are not logged
and which ones are logged incorrectly). By the way, there's room for
implementing other checks.

Which abbreviated hash length the bot should enforce is
debatable. Currently, it suggests the sha length of 8, and I was
thinking about increasing it to 12. In the end I decided not to, but
here was my logic:

* From 2016.11 to 2017.04.2, seven characters were used.
* 2017.04.3 had ten character shas in the changelog.
* From 2017.05 to 2017.07, shas had eight characters.
* If you ask git to give you an abbreviated hash, it gives
  nine characters.
* It makes little sense to use abbreviated hashes that are smaller
  than what git produces.
* Therefore, let's bump it to 9 characters…
* However, git project repo already has collisions with 8 characters
  and linux repo has collisions with 9, so in a long run even 9 is not
  that good of an idea…
* So let's use 12.
* But even 12 may at some point may stop being sufficient (it's
  never enough).
* Instead, let's keep it at 8. Once we start approaching birthday
  collisions for some length, we can bump it in the whole changelog.

And there you go! Yet another blog post in a commit message 😉