More aggressive parallelization of tests
Now if the bot is blocked by something else, it can send SIGUSR1 to
the test bench to extend the timeout. This way bots can wait forever
for the build to become unlocked.

Existing timeouts were a bit too tight, so with the load of 8 bots
being tested, some of the tests just can't make it in time. Therefore,
timeouts were extended (1.5x) under ｢prove｣.

See issue #205. Now we are down to 4m 32s from original 22m.