Markdownify quotable output
Issue #152. Also, the script used to fetch IRC logs is now finally
committed.

Now it also processes every channel in parallel, so it is a little bit
faster. That said, the grep itself became slower because of the extra
code in it. Overall there's almost no change in speed.

Related: issue #27