Add Regressionable
Issue #123.

This is not a real bot, and in fact I'm doubtful that it is useful at
all. Even though I successfully used it to find a bunch of
regressions, I don't think it should be part of the whateverable
project.

In any case, for now let's have it committed somewhere.