More aggressive compression
This saves about 21G of storage (≈⅓ shaved off). The midpoint is
supposed to be in front of the cutoff point so that bisectable
stumbles about the fast build at its first step. I haven't counted the
actual number of commits, but it should be right. Even if not, it'll
fix itself in a month or so.