⚠ Make all bots runnable locally (this is so cool!)
Previously if you wanted to run bisectable or some other bot locally
you had to build tens (if not thousands) of rakudo versions to get
something useful out of it. Now whenever it stumbles upon a missing
build it will attempt to pull it from the main server. If you're on
linux x86_64, this means that the bots you start are no worse than the
ones running on the main server.

Note, however, that this does not let you *use* these bots as tools
locally. Well, unless you fire up your own IRC server and configure
the thing to use it. We will get this working next, for now this is a
huge step already (see #122).

Ah, almost forgot to mention, this does not work. See
https://github.com/rakudo/rakudo/issues/1501. I mean, yeah, you can
probably use committable, evalable and some other bots, but bisectable
will run out of memory before you get to any useful result.