Time out later if we're downloading builds
Most tests have a timeout, and some of these tests need tens of rakudo
builds (which we have to download if we're running locally). With this
change it will ask the test suite to wait some extra time which is
roughly equal to the amount of time we spend on fetching the builds.