Support older zstd version, complain if not installed
Tackles issue #122 a little bit. This *should* make it work for older
zstd versions like 0.5.1 (but I haven't tested it just yet).

Also, it should now die with a more helpful message if zstd or lrzip
is not installed.

I think this adds about 20ms delay every time the code attempts to use
some build, and that may seem negligible but it's not. Bisectable
normally needs ≈13 steps, which is a ≈0.26s delay. Committable
currently uses 28 commits for “releases” or “v6.c” query, and that is
a 0.56s delay. Even small things add up in this case, so we can
declare this path kinda hot and add some caching… See issue #294.