Tickets marked `testneeded` are not just updated
Fresh tickets on RT are “new”. When someone marks a “new” ticket with
`testneeded` tag and writes a comment, the state changes to
“open”. For that reason reportable was showing some tickets as
[new→open] instead of putting them into “half-resolved” category. This
is now fixed.

This means in some relatively rare cases reportable used to generate
LTA reports. They were not fully wrong, but they could've been
righter. You can just ask reportable to generate a new report for any
date in the past and it will produce the right output now.