Fix license field in META6.json
According to https://spdx.org/licenses/ “AGPL-3.0” identifier is
deprecated, and for a good reason. Changed it to “AGPL-3.0-or-later”.

Note that this is not relicensing, the code already comes with “or
later” clause in a header of every source file. It is merely making
META6.json file consistent with the reality.