Handle failures in promised results
We attempt to generate a useful response when we get an unhandled
exception. This was true for any irc-to-me except for those that
returned a Promise. Now it will handle this case also. This is useful
because long-running jobs can cause bots to ping out, and now it is
possible to just wrap anything in a start block.