Add :wipe named parameter in run-smth
As well as another helper method to avoid run-smth altogether. This
allows heavy parallelization of things that need to use the same
build at the same time. Quotable, Benchable and Project Blin can
benefit from that. Note that if you use that you'll have to remove
unpacked builds yourself (the location of them is returned by
run-smth-build-path).