Avoid LAST to work around some weird issue
Sometimes when run-bisect runs from multiple threads at the same time
I am getting Nil in string context warnings. To me that can only mean
that `$result<output>` is not what it is supposed to be, so there's a
bug in rakudo. Working around it because I want to move on right
now. Please feel free to investigate.