Implement user list and automatic .tell
Previously evalable6 and unicodable6 had to run NAMES command in order
to see if there's a specific user on the channel. Now bots
automatically keep track of join/part/quit/nick events, relying on
NAMES only to get the initial list. They will also use NAMES to
refresh the list from time to time, which is somehow needed in my IRC
client during netsplits and such.

Also, thanks to this tellable6 now implements automatic .tell when
you try to speak to a real user who is no longer on the channel.