Evalable: respond to discord even if camelia is present
Normally evalable6 leaves queries to camelia when she is around.
But camelia does not support responding to discord, so in that case
evalable6 should not restrain itself.

This is made possible by mixing a FromDiscord role into the nick
field of messages (not the message itself because that would break
event dispatch in IRC::Client) and testing for that while deciding
whether to leave the message to camelia or not.