Make Sourceable more awesome
TL;DR: You can just give it the code you have as is. Magic.

The best interface is no interface. Previous version was a simple
cover for the CoreHackers::Sourcery module, which is a collection of
helper functions to get .file and .line of code blocks. It was OK,
except that it required users to transform their code to a particular
format in order to get the right output. This is no longer
needed. With this rewrite, the original code as well as the dependency
on CoreHackers::Sourcery are completely removed. Instead, the code
implements whateverable-style crazy heuristics/brute force that tends
to Just Work™ in pretty much all cases. To prove this, I used queries
(found in the IRC log) to the previous version of the bot
(SourceBaby), and was unable to find any case in which the current
implementation did not work. Of course, creative users will eventually
stumble upon something, so we'll likely need to improve it later.

Also, now it's possible to use it for any rakudo commit. Of course.