Make Bisectable even more awesome
Previously it was assumed that the user knows that there was in fact a
change in behavior before triggering the bot. It made sense because
there was a separate bot (Committable) that allowed to get that
information. However, this expectation doesn't match reality and how
actual users are using the bot.

This commit combines functionality of Committable and Bisectable into
one, meaning that by default you can just give a piece of code to
Bisectable and it will first get the output on all releases and only
then bisect all of the changes if possible/needed. Sometimes no
interface is better than a good interface, and a single “do something”
command is better than requiring to run multiple commands and
understanding what all of them do and why.

Previous behavior is preserved as long as you give any arguments (like
old= or new=).

The output is currently somewhat noisy but we will see how it goes. If
it turns out to be more annoying than useful then something will be
done about it.

Somewhat related: #128 #257