New bot: Buildable
For now this is simply a rename of the long-term-storage.p6 script. It
used to be that build.p6 was started by a cron job and
long-term-storage.p6 was started manually. Because ideally the
building should start immediately after a commit was pushed, there
needs to be a persistent process. For now this will be a separate bot,
but in the future I might merge it with shareable or something else.