Switch the IRC host to libera
Because freenode is no longer a thing (and everything moved to libera).
Note that this tweak was hastily applied on the server and remained
uncommitted, so actually this commit changes nothing, it's just about
committing what is already there.

The IP address in this commit is what I got when pinging
`irc.libera.chat`, it's needed to work around the issue of having too
many bots per IP (connecting to an IPv4 address forcefully uses IPv4
address, which is treated like a completely different connection and
thus has its own limit). It's a very old hack that for now has to
remain, I think. And yes, that IP can change over time, but until there
is a better fix I'm not sure if anything can be done about it.