Improve Tellable's .seen normalization on startup
Tellable stores a dict of all latest messages by all users that it has
seen. However, the keys in that dict are not exact nicknames. Instead,
they are intentionally mangled simplified versions of nicknames. The
goal of the normalization is to allow users to change their nicknames
slightly and still receive messages, and have `.seen` that DWIMs.
AlexDaniel, AlexDaniel11, alexdaniel_, etc. are all treated the same by
Tellable, that's what makes this bot so nice. 😌

Now, when the bot starts up, it could be that the file was created with
old rules, or the file was recovered from something that didn't follow
the same rules. For that reason the code renormalizes the entire seen
and tell files and saves them back. However, now that I looked at the
code again, I realized that there is a flaw. If two nicknames get
normalized to the same value, it should pick the latest message and
discard the old one. However, the code was just keeping one random
message instead of taking the timestamp into account. This commit fixes
that.