Fix Buildable webhook endopoint
GitHub is making POST requests, not just GET. Note that in this case
we're just interested in the push event itself, not about the data, so
there are no secrets or anything else, and it's fine. The endpoint is
public and everyone is free to trigger the builds if they want
to (though why would they do that?).