Stop Buildable from pretending to work when asked
The idea is that if any other mechanism breaks (e.g. webhooks are down),
then you should be able to “wake” it up by just asking for the status.
However, in this case the functionality was a bit too eager, which
resulted in the bot always saying “Building” even when it was idle.