Fix `:cwd` thinko from months ago
I'm not entirely sure how this is working in the currently running bots,
but this is definitely causing issues in Blin. Basically, when the
directory structure was changed, I changed these lines incorrectly. As
the result, when running this sub, it was trying to `:cwd` into the URL
of the repo (which is obviously wrong).